moved contact handling to contact context
